targets to device

+++ src/zoc/baseline.py
@@ -151,7 +151,7 @@ def train_id_classifier(train_set, eval_set):
         # train
         for image_features, targets in tqdm(train_loader):
             image_features = image_features.to(torch.float32).to(device)
-            targets = targets.to(torch.float32).to(device)
+            targets = targets.to(device)
 
             optimizer.zero_grad()
 
@@ -174,7 +174,7 @@ def train_id_classifier(train_set, eval_set):
         for eval_features, eval_targets in tqdm(eval_loader):
 
             eval_features = eval_features.to(torch.float32).to(device)
-            eval_targets = eval_targets.to(torch.float32).to(device)
+            eval_targets = eval_targets.to(device)
 
             with torch.no_grad():
                 eval_preds = classifier(eval_features)
